daily_expenses/requirements.txt
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.8.1
autopep8==1.7.0
bleach==5.0.1
cffi==1.15.1
comm==0.1.2
debugpy==1.6.5
defusedxml==0.7.1
Django==5.1.2
djangorestframework==3.15.2
entrypoints==0.4
fastjsonschema==2.16.2
fqdn==1.5.1
ipykernel==6.19.4
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-events==0.5.0
jupyter_client==7.4.8
jupyter_core==5.1.2
jupyter_server==2.0.6
jupyter_server_terminals==0.4.3
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
mistune==2.0.4
mysqlclient==2.2.5
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.7
nbformat==5.7.1
nest-asyncio==1.5.6
notebook==6.5.2
notebook_shim==0.2.2
pandocfilters==1.5.0
platformdirs==2.6.2
prometheus-client==0.15.0
psutil==5.9.4
pycodestyle==2.9.1
pycparser==2.21
PyMySQL==1.1.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-json-logger==2.0.4
pyzmq==24.0.1
qtconsole==5.4.0
QtPy==2.3.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
Send2Trash==1.8.0
sniffio==1.3.0
sqlparse==0.5.1
terminado==0.17.1
tinycss2==1.2.1
toml==0.10.2
tornado==6.2
traitlets==5.8.0
typing_extensions==4.12.2
uri-template==1.2.0
webcolors==1.12
websocket-client==1.4.2
widgetsnbextension==4.0.5
